Random Disconnection Issue Fixed
Random disconnection issue was caused by a missed interrupt by the MAX30003. The source of this issue was the fact that we are logging information in line.

Logging has been moved to when the processor has time. The log buffer has been increased.

On production level boards, logging has been disabled.